GCP: conditionally create bootstrap service acct
This changes the default behavior of always generating a service account
to sign the ignition URL, to only generating the service account when
it is needed; i.e. when authenticating with environmental
authentication. Most of the time users provide a service account with
key so we can just use that, as we did before environmental auth was
recently introduced.