OCPBUGS-7495: Convert platform type for AgentClusterInstall
The platform types in AgentClusterInstall are in mixed case. Previously
we were just casting the lowercase platform name to set the PlatformType
field. This resulted in the platform type being silently ignored when
creating the Cluster in the assisted API. Clusters were getting the
default type for their topology (None for SNO, otherwise BareMetal), but
there was no way to set VSphere as the platform.