asset: azure: add `Plan` field to OSImage
Marketplace images were hardcoded to always require a purchase plan.
However, some use cases like ARO offer images that do not require a
purchase plan.

This change adds an option to the OSImage field in the installconfig to
specify images that do not require purchase plan. If omitted, it
defaults to the current behaviour of assuming
AzureImageTypeMarketplaceWithPlan.