manifests: Fix file check for loading openshift manifests
When generating the openshift manifests, it tries to load the
dependency control plane asset that creates files in the
openshift directory. It then checks to see if the openshift
manifest is actually already generated before this by checking
the openshift directory for files that are not generated by the
machine asset.

The newly created CPMS file was missing from the check, hence
the asset loader always thought the asset was generated since
there are files that exist other than the machine files and exits
but it actually had to be newly generated with 4 files that were
crucial to installation.

Adding the CPMS files to the list of machine asset files.