azure: session: fix certificate json annotations
The `autorest/auth` library expects specific keys in the json file for
certificate authentication [1]. If we ever saved the Credentials object
to a json file during survey, it would have used the wrong keys:
`certificatePath` and `certificatePassword` instead of
`clientCertificate` and `clientCertificatePassword`, respectively.

This change aims to fix that so that the saved `osServicePrincipal.json`
file with certificate authentication details can be correctly loaded by
the autorest library. It'll also be important in a follow-up change
where we are going to load the file ourselves instead of using autorest.

Notice that this is not breaking compatibility with previous versions
since those fields were never used to create a file (we only survey
client secret credentials).

[1] https://github.com/Azure/go-autorest/blob/main/autorest/azure/auth/auth.go#L348-L349