OCPBUGS-8094: In agent 'wait-for bootstrap' command, test ssh to Node0
If the Rest API and Kube API are not reachable, it may be because
network connectivity checks are preventing the install from progressing
(which will also prevent the Node0 SSH server from starting). Attempt
to SSH to the node and provide instructions for further debug to the
user upon failure.