Azure: Add outbound pool for public load balancer
When a node is removed from the public load balancer, if it is not
attached to an internal load balancer, it falls back to the default
outbound connectivity which permits traffic. If it is attached to an
internal load balancer however, outbound connectivity is blocked until
the VM is re-attached to a public load balancer, public IP, or a NAT
gateway is added to the subnet.

This solution adds a secondary public IP address and backend pool with
the control plane machines onto the public load balancer.

https://issues.redhat.com/browse/OCPBUGS-7359