bootstrap-pivot: skip pivot in SCOS Live ISO
If bootstrap is running in a Live ISO of SCOS `rpm-ostree pivot` will
fail as /usr is not writable. Live ISO is used in assisted-installer
flow and it doesn't need to pivot machine-os-content during bootstrap
(it would write SCOS image on disk and run pivot via
machine-config-daemon-pivot service). This change unblocks OKD Stream
isntall using assisted-installer