gcp: add confidential compute support for boostrap TF
The confidential compute configuration options for the bootstrap
instance are defined by the respective control-plane instance options.

The scheduling and confidential_instance_config options are defined as
dynamic blocks and they will be applied when the user specifies the
respective config options. This allows us to respect the cloud provider
defaults, when the user does not provide any values. At the same
time, it does not require code updates when the cloud provider defaults
change.

Signed-off-by: Michail Resvanis <mresvani@redhat.com>