MGMT-13793: Do not remove host default configuration unless network configuration is provided for it
Presently, in the PreNetworkConfig script, we are removing the default network manager configuration for a host in every case. This should be altered so that the default configuration is deleted only if custom nmstate configuration has been provided for the host.

This PR also addresses a breaking issue with the way in which Dracut Initqueue scripts are handled. Consider the following directory on a given host.

ls /usr/lib/dracut/hooks/initqueue/settled
90-assisted-pre-static-network-config.sh 99-nm-run.sh

if we inject a script between the two scripts with the sole purpose of using an exit statement...

echo "exit 0" > /usr/lib/dracut/hooks/initqueue/settled/91-test-exit.sh && chmod +x /usr/lib/dracut/hooks/initqueue/settled/91-test-exit.sh

Then the script 99-nm-run.sh will not be executed. It appears that an exit in any of the .sh scripts is considered to be a "fatal" error.

So the best practice is to avoid using exit for "non fatal" situations. The code in this PR has been altered to reflect this.