OCPBUGS-10342: Check that number of replicas matches hosts
If the number of compute replicas is not configured, a default
of 3 will be used. If the number of baremetal hosts configured
does not match this, a failure will occur at installation time.

Add a couple messages when the ISO is created to help catch
config problems with the replicas before installation.
- An info message with the number of replicas being used which is
useful if not configured
- A warning message if the number of replicas does not match
the number of configured hosts