Remove dead api_vip setting code
This has not been called since 4b0a221641a4efce1b443d2a2fe5240f66acd6e8
accidentally introduced a check on the (non-existent) environment
variable ${APIVIP} instead of the template variable {{.APIVIP}}. Since
that hasn't seemed to be a problem, we can remove it altogether.

The code to set the API VIP was originally introduced waaay back in
4b0a221641a4efce1b443d2a2fe5240f66acd6e8. It was there to work around an
issue that was solved a couple of days later by
openshift/assisted-service#3258.