CORS-2525: Azure: remove storage account with bootstrap destroy
The work is being done to address the issue of Azure storage accounts being persisted for the life of an OpenShift cluster, which incurs costs and management effort. The goal is to destroy the storage account along with other bootstrap resources during the bootstrap destroy process. Additionally, the default creation of two Azure storage accounts has raised security issues, which the work aims to address by restricting network access using virtual network rules. The end result is a more cost-effective and secure OpenShift cluster deployment on Azure.