OSASINFRA-3111: Warn if user/pass auth used in clouds.yaml (#6911)
Application credentials [1] have a number of advantage over
username/password combinations, not least the ease of rotation. However,
the 'clouds.yaml' files generated by various OpenStack tools like
Horizon default to username/password combinations and users tend to pass
these through unmodified. We don't want to outright reject these - at
least, not yet - but we can indicate an explicit preference for
application credentials. We do this by inspecting the 'clouds.yaml' file
that the user has provided before we store it as a secret for use in the
deployment. If the user has provided password-based credentials, we will
emit a log encouraging them to migrate and warning about the potential
removal of support for password-based auth in a future release.

[1] https://docs.openstack.org/keystone/latest/user/application_credentials.html

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>