azure: session: enable Managed Identity authentication
Add support for both Managed Identity and User-assigned identities. The
latter is selected by the presence of `ClientID` in the
`osServicePrincipal.json` credentials file while the former is selected
as the last auth type if no other could be found.

osServicePrincipal.json for user-assigned identity:
```
{"subscriptionId":"<Redacted>","tenantId":"<Redacted>","clientId":"<User-assigned
identity ID>"}
```

osServicePrincipal.json for system managed identity:
```
{"subscriptionId":"<Redacted>","tenantId":"<Redacted>"}
```