SPLAT-1034: Allow CIDR expansion on single zone deployment in AWS
The goal of this change is to provide flexibility for subnet expansion
when deploying a cluster in AWS with default IPI in single-zone on AWS.

Currently the installer tries a best-effort to fill all the CIDR blocks
from VPC spliting into two blocks: public and private. The problem when
creating in single-zone, it could prevent VPC expansion as it will create
two subnets /17, when using MachineCIDR (VPC) with /16.

This change checks the length of zones, and when the user-provided only
one Availability Zone on install-config.yaml, it will increase one bit
on the available blocks for public and private, leaving free at least
50% of the VPC for future expansions, like creating manually subnets
into different locations (Local Zones) when the user need it. It also
allows the user to create a cluster cheaper then expand it without needing
to create a new cluster with new configuration.

Examples:

- Using MachineCIDR /16, two subnets /18 is created, with empty
  block /17.
- Using MachineCIDR /20, two subnets /22 created, with empty
  block /21.

Use case:

- As a User I would like to evaluate OCP minimally in AWS, single zone,
  then be able to increase to more availability zones
- As a User extending nodes in Local Zones, I would like to create new
  subnets on recently announced AWS Local Zones locations, and MachineSet
  manifests, so I can extend my nodes, and workloads, to the metropolitan regions
  in Local Zones. (supported on 4.12+)

References:

- https://issues.redhat.com/browse/SPLAT-1034
- https://issues.redhat.com/browse/SPLAT-657