images: decouple installer and installer-artifacts
The installer-artifacts image depends on the installer image,
which means they cannot be built in parallel. The effect of this
dependency is that the installer-artifacts image contains the linux
amd64 builds in addition to the mac/linux amd64/arm64 builds present
in the installer-artifacts image.

Breaking the dependency results in a more efficient build process
because they can be built in parallel as well as no duplication of
the linux amd64 binary in the release image (before this change the
binary would be present in both images within the release image).

The risk, though, is that clients are inconsistently grabbing the linux
amd64 binary from either image, rather than only the installer image,
in which case we would break any client trying to get that binary from
the installer-artifacts image.