vSphere: terraform / machine changes for template
This commit adds terraform variable changes, machine
and machine set changes and validation for the existing
virtual machine rhcos template.

Confirm that the template exists in vSphere
Confirm that the rhcos version is within one
release behind current installer.
If the image is ahead of current installer fail.
If the image is behind the installer 2 or more releases fail
If the image is 1 release behind warn.