vSphere: add ova sha query; additional debugging
We have been experiencing issues with OVA
corruption for quite a while. In investigating
this issue it seems the provide sha256 for the OVA
was being left unused. This PR adds `sha256` query
so that upon download of the OVA the integrity
will be checked.

In addition the content length of the body GET
along with the io.Copy bytes written is outputted
when debugging is enabled.

Further debugging was added to the vsphereprivate
resource if and when parsing of the OVF descriptor
fails. It checks the SHA256 and provides the size of
the OVA for additional checking.