AGENT-562: Load config from config image
Add a service to run a script to load the contents of the archive created
in AGENT-563 to the system. The script will check for a device with
the config image label and copy its contents.

This service will eventually be added to the unconfigured ignition
as part of AGENT-558.