aws tf: add terraform automation to create subnets in Local Zones
Adding terraform automation to create resources subnet and route
table associations in AWS Local Zones.

The AWS Local Zones are identified by edge zones as it can be
expanded to used with Wavelength Zones which uses similar
technology. Wavelength Zones is not coverred by this feature.

The Local Zones does not support Nat Gateway, for that reason
the terraform will create only subnet and associations to the
route table from the parent region, when exists, otherwise
the first route table will be created.

The subnets in Local Zones will be created only when the zones
is supplied on the install-config.yaml in the edge compute pool.

The CIDR blocks is also changed to support this feature. When
there are Local Zones, the available CIDR provided to Machine
network (VPC) will be sliced into more blocks. It must be done
to differentiate the regular zones from edge zones, and prevent
trying to create unsupported resources like Nat Gateways.

The change on CIDR blocks was done to prioritize the private
subnets, slicing only public subnet blocks to Local Zone subnets.

To prevent CIDR blocks exhaustion, and blocking subnet expansion
in Day 2, the terraform will detect if the installation is using
single zone and preventing to use all the block for a single subnet.