aws tf: terraform automation to create subnets in Wavelength Zones
Adding terraform automation to create resources subnet and route
table associations in AWS Wavelength zones.

The AWS Wavelength Zones are identified as edge zones by installer.

The Wavelength Zones does not support Nat Gateway, for that reason
the terraform will create only subnet and associations to the
route table from the parent region, when exists, otherwise
the first private route table will be used in the association.

The subnets in Wavelength Zones will be created only when the zone
names are supplied in the install-config.yaml in the edge compute pool.

AWS Wavelength requires an different type of gateway when ingress/egress
traffic from the zone: Carrier Gateway.

When installer creates the VPC, the terraform creates:
- the Carrier Gateway associating to the VPC
- public edge route table
- public subnet in Wavelength Zone, associating to the public edge route
  table

The installer odes not create Machine Set configuration to launch edge
nodes to public subnets, but the user can do it in install time, for
that reason an different feature is required to support MAPI AWS
Provider, not covered and not blockes the full automation delivered by
installer.