aws tfvars: create vars for Local Zones, source and edge gateway map
This change export all the available AWS zones to be used as a source
when creating the terraform variables.

The zones are used to create a 'gateway mapping' to identify the
Local Zone's parent zone, and create a map to terraform identify
the route table ID from the parent zone, when exists, then associate
the Local Zone private subnet. This change must be done for two reasons:
1. Local Zone does not support Nat Gateways, so the gateway used will be
the parent zone.
2. to keep deliver the best pratices of creating public and private
subnets in installer automation, placing the nodes in Local Zones in
the private subnets.

This change also select the edge zones from the regular zones in compute
pool aggregated by source `WorkerConfigs`.

The ZoneGroup is also selected, the to further be used by terraform
to opt-in the Local Zone location. Note: each location must be opt-in before
creating any resource on that, after that it will be available for the
account, it must be done one time.