aws validation: install-config for edge pool to new subnets
The edge pool is valid for AWS Local Zones when installing a
cluster in existing subnets and, now, in new clusters.

The edge compute pool in install-config.yaml requires the platform
specific zone field when the subnets is not provided, assuming
a new cluster, so the installer will create the subnets into those
zones.