aws cfg meta: add support to store zone attributes
The AWS Zone provides a few attributes that is essential for the
edge compute pool when setting up AWS Local Zones.

The `GroupName` is used to create the labels that can be used
by user-workloads when deploying applications into specific locations.

The `Type` is used to aggregate edge and regular worker zones. Only
`local-zone` type is valid for edge compute pool. This field also
can be used on the user-workloads and it is exposed through Machine/node
label.

The `ParentZoneName` the zone in the region that the Local Zone is attached to.
It is used by installer to discover which VPC private route table the private
subnets on the Local Zones will be associated as t Local Zones does not
support Nat Gateways (and other network resources).

This change create the structure to store the zone attribute to installer
be used into different places (MachineSet creation, EC2 offerings, Gateway
map for tfvars, etc).

It also introduce functions used to discover and filter the zones in the
region.