aws compute pool: add suport to new subnets for local-zones
Support to create MachineSet manifests in new installations,
when only the zones is supplied in the edge compute pool
for Local Zones (new VPC).

In that installation type (new VPC), the subnets structure,
used to create MachineSets, is generated based on the zones
provided in the `compute[?name=="edge"].platform.aws.zones`.

The Instance Type used will depend of the zone offerings when
querying the AWS EC2 Instance Offerings API. The instance
is selected based on the 'allowed/preferred' list from installer.

The label `machine.openshift.io/parent-zone-name` is also added
to the MachineSet manifest. It is not used anywhere than as a
helper for user-workloads creates affinity with applications
running in Local Zones and in the Availability Zones.