baremetal bootstrap: Stop ironic in master-bmh-update, not startironic
This code at the end on startironic.sh to shut down all of the ironic
containers is apparently needed in order to release the API VIP after
provisioning has finished.

The code to get the current status is duplicated from master-bmh-update,
and is waiting on that script to finish in order to trigger, so move the
container shutdown code there.