baremetal bootstrap: Rename env vars to match containers
Don't use different names for environment variables from what we need to
pass to containers. This enhances readability and makes mistakes less
likely.

Note that the rename of IRONIC_KERNEL_PARAMS means that the additional
rd.net.timeout.carrier option from
https://issues.redhat.com/browse/OCPBUGS-5151 is now only passed to
ironic and ironic-inspector, *not* to coreos-downloader. This should
give the same behaviour for the control plane as the equivalent fix for
workers in https://issues.redhat.com/browse/OCPBUGS-872. The IP_OPTIONS
environment variable in machine-os-images only affects the ISO, which is
only used in virtualmedia deployments.