baremetal: Use static libvirt library to destroy cluster
Currently we link to a C-based dynamic library (via the libvirt-go
module) to connect to libvirt for the purposes of destroying a baremetal
cluster - which in fact only deletes the storage pools used by the
bootstrap (and is only used to clean up in cases where the installation
fails, since otherwise terraform will clean up when the bootstrap is
torn down).

The Terraform provider that actually creates the bootstrap uses a
pure-Golang library (go-libvirt). By switching to that same library here
we can avoid the need to dynamically link to libvirt.