aws: validate hostedZoneRole cred mode
Adds validation to require that either manual or
passthrough credentials mode is used when specifying
hosted zone role.

In order to perform the AssumeRole operation on the provided
role, a policy must be in place to establish a trust reltionship
between the role and the IAM credential (for the installer and
ingress operator). Because mint mode will create new credentials
with non-deterministic unique identifiers in the cluster,
it is impossible to generate the policy in advance.