OCPBUGS-14877: Validate that number hosts does not exceed replicas
A check was added in https://issues.redhat.com//browse/OCPBUGS-10342
to warn when the number of replicas exceeded the configured hosts,
however this did not catch the case when the number of configured
hosts exceeds the replicas, so it is added here. In addition this
validation will now result in an error instead of a warning since
this invalid configuration will cause installation failures.