aws: Make bootstrap S3 bucket optional during bootstrap destroy
Making a copy of the AWS bootstrap S3 bucket in the cluster stage
and adding a field to the install config. If the user does not want
to delete the S3 bucket created for the bootstrap, they need to
set the new field to true.

This causes the S3 bucket to be created during the cluster stage
and not the boostrap stage which will then be ignored when the
bootstrap destroy code runs.

More details: https://issues.redhat.com/browse/CORS-2661