OCPBUGS-15861: Load libvirt library on demand
The libvirt-go library can now dynamically load the libvirt client
sharted library on demand using dlopen(), instead of linking to it
dynamically (which requires the shared library to always be present at
both build and run time).

This means that the same Linux binaries can be used even on systems that
don't have the libvirt client installed, as long as they are needed only
for code that doesn't involve calling libvirt. (Currently this is
everything except the baremetal platform and the 'destroy cluster'
command on the libvirt platform. All other interaction with libvirt is
via Terraform.)