ic: gcp: improve instance type validation
The instance types were just being validated if they were explicitly set.
However that's not enough since controlPlane and compute nodes can
"inherit" the instance type from the defaultMachinePlatform or an
Installer default if nothing is set.

This change properly derives the instance type and makes sure it is
validated in terms of minimum requirements and user-configured zones.