GCP XPN: clarify service account support
Clarify the wording in the service account field in the gcp machine
pool to indicate the field is only supported in the control plane
machine set.

Because compute and control plane machine pools are the same type,
openshift-install explain will display the service account field
in the compute machine pool. Hopefully this will make it slightly
less confusing.