OCPBUGS-16204: aws: attach additional security groups to controlPlane
The security groups are being properly added to the controlPlane
resource definition manifest but they are not being piped through to
terraform to be attached to the instances.

This fixes the issue that, after installation, checking the Security
Groups attached to master and worker, master doesn’t have the specified
custom security groups attached while workers have.

For one of the masters:
```
[root@preserve-gpei-worker ~]# aws ec2 describe-instances --instance-ids i-0cd007cca57c86ee9 --region us-west-2 --query 'Reservations[*].Instances[*].SecurityGroups[*]' --output json
[
    [
        [
            {
                "GroupName": "terraform-20230713031140984600000002",
                "GroupId": "sg-05495718555950f77"
            }
        ]
    ]
]
```