aws edge lookup: improve warnings when type not found
The installer discover the offerings by zone from a supported set of instance
types, when none of those are available in the zone it emits a warning.

Edge compute pool diferentiates from regular zones when selecting the
EC2 Types, it may not share the same type by edge zone, instead every
zone type could have a different EC2 type, depending of the offerings in
the zone.

The warning for edge zone shows which zone does not have any in the
supported list, to help the user to update the machine set manifest in
later stage.