OpenStack: Remove SGS created by CPO on destroy
cloud-provider-openstack can be configured to create security groups for
the NodePorts of the load balancers it is creating. These SGs are then
attached to the nodes. On `cluster destroy` we're orphaning them. This
commit makes sure that we're looking for them.

As they aren't tagged or have a proper cluster ID in the name, we will
look at each of the ports, list its SGs and evaluate them comparing
names with the pattern. If it matches, `destroy` will attempt to delete
such SG.