OCPBUGS-17227: gcp: fix validation of custom instance types
Users can create their own instance types in GCloud. Those types are not
returned when listing the machine types for a given project/region. This
would cause those installs with custom types to fail with the error:

```
Internal error: failed to fetch instance type, this error usually occurs if the region or the instance type is not found
```

after the validations introduced by
https://github.com/openshift/installer/pull/7317 and
https://github.com/openshift/installer/pull/7096.

To fix that, let's fallback to the previous way of fetching the specific
machine type whenever the aggregated list returns 0 elements.