OpenStack: Set external network for cloud-provider
We allow users to specify an external network to create the floating IPs
from on the install-config.yaml. We use that to create the API and
Ingress FIPs and forget about it. We can however also use it to
configure the cloud-provider-openstack to make sure it'll choose the
correct network when creating FIPs for the `type=LoadBalancer` Services
too. This commit implements it.