OCPBUGS-7690: azure: destroy: dns records leak if permissions missing
If the Service Principal destroying an Azure cluster lacks the
permissions to list DNS records, the Installer will complete the destroy
but leave DNS records behind.

Unfortunately the Azure SDK doesn't return any errors when permissions
are missing neither does it offer a way for us to check if the required
permissions are present. As an in-between solution, let's log a warning
when we can't find any DNS records so users know to check for any
records being leaked.