OCPBUGS-17861: cinder volume type/az validation
For each rootVolume type specified that has zones configured in Cinder,
check that the zones were specified in the machine pool with the rootVolume zones field.

See https://docs.openstack.org/cinder/latest/admin/availability-zone-type.html for more information.

Note that this is not a complete check, as if more than one type is specified,
the zones could be specified in another index of the rootVolume zones field.
For that we should add some validation that zones found in a type at a
specific index, matches the zone name of that index in root volume zones
field.