Inline staticnetworkconfig from assisted-service
In later versions of assisted-service, this code changes from shelling
out to nmstatectl and instead uses the libnmstate library. Since we
don't have a way of building with that yet, this is preventing us from
updating assisted-service.

Move the nmstatectl-based implementation in-tree so we can continue to
use it independently of assisted-service.