OCPBUGS-2339: Restart agent.service if assisted-service restarts
The agent service is unable to continue to next step if
assisted-service restarts and clears its internal database.

A new service named agent-restart.service now checks if
assisted-service goes offline. When it goes offline, the new
service restarts agent.service and itself. This forces the agent
to reregister with assisted-service and allows installation
to continue.

This fix does not handle the situation where the cluster has
started installation and assisted-service is then restarted.
In that case, all hosts will need to be rebooted.