Don't depend on misspelling to default userManagedNetworking
The code to set the default value for userManagedNetworking was inside
some backwards compatibility code that converted the erroneous platform
name "none" to the correct "None" in AgentClusterInstall. This meant
that it only ran when the user supplied the wrong name.

Move the code outside so that it runs whenever the platform name ends up
as "None", regardless of whether that was correctly specified by the
user or translated for backwards compat.