OCPBUGS-18428: Add ip=dhcp,dhcp6 kernel param for vSphere dual-stack
We have observed that when booting up the node, IPv4 and IPv6 stacks can
come up at different times causing OpenShift networking services to
start when IPv4 is ready but IPv6 is not. This is not desired for
dual-stack setups where both stacks should be up before we proceed
further.

With this PR we are adding `ip=dhcp,dhcp6` kernel param to master and
worker nodes in vSphere scenario when cluster is deployed as dual-stack.

This PR handles only a scenario of initial cluster installation.
Conversion from single-stack to dual-stack will be handled separately.

Fixes: OCPBUGS-18428
Contributes-to: OPNET-230